requirements.txt
pandas
yfinance
SQLAlchemy
python-dotenv
openpyxl
requests